rename resolve util to module
