+++ code/core/src/common/utils/get-addon-annotations.ts
@@ -1,4 +1,4 @@
-import { resolveModule } from '../../shared/utils/resolve';
+import { resolveModule } from '../../shared/utils/module';
 import { isCorePackage } from './cli';
 
 /**
