+++ code/core/src/common/utils/interpret-files.ts
@@ -1,21 +1,24 @@
 import { existsSync } from 'node:fs';
 
-export const boost = new Set(['.js', '.jsx', '.ts', '.tsx', '.cts', '.mts', '.cjs', '.mjs']);
-
-function sortExtensions() {
-  return [...Array.from(boost)];
-}
-
-const possibleExtensions = sortExtensions();
+export const supportedExtensions = [
+  '.js',
+  '.mjs',
+  '.cjs',
+  '.jsx',
+  '.ts',
+  '.mts',
+  '.cts',
+  '.tsx',
+] as const;
 
 export function getInterpretedFile(pathToFile: string) {
-  return possibleExtensions
+  return supportedExtensions
     .map((ext) => (pathToFile.endsWith(ext) ? pathToFile : `${pathToFile}${ext}`))
     .find((candidate) => existsSync(candidate));
 }
 
 export function getInterpretedFileWithExt(pathToFile: string) {
-  return possibleExtensions
+  return supportedExtensions
     .map((ext) => ({ path: pathToFile.endsWith(ext) ? pathToFile : `${pathToFile}${ext}`, ext }))
     .find((candidate) => existsSync(candidate.path));
 }
