+++ code/core/src/common/utils/validate-configuration-files.ts
@@ -8,10 +8,10 @@ import { glob } from 'glob';
 import slash from 'slash';
 import { dedent } from 'ts-dedent';
 
-import { boost } from './interpret-files';
+import { supportedExtensions } from './interpret-files';
 
 export async function validateConfigurationFiles(configDir: string, cwd?: string) {
-  const extensionsPattern = `{${Array.from(boost).join(',')}}`;
+  const extensionsPattern = `{${Array.from(supportedExtensions).join(',')}}`;
   const mainConfigMatches = await glob(slash(resolve(configDir, `main${extensionsPattern}`)), {
     cwd: cwd ?? process.cwd(),
   });
